fix(ios): surface Delete Account in the account menu
Apple 5.1.1(v) requires account deletion to be easily discoverable. The
Delete Account entry only lived on the /profile page, which on iPhone is
buried in the bottom-bar More overflow and on iPad is unreachable (the
profile nav entry is only added in row/mobile layout). The avatar
dropdown the user naturally opens (Settings/Distribution/Console/Log Out)
had no delete entry at all.

Extract the delete flow into a shared DeleteAccountDialog component and
add a Delete Account item to the avatar dropdown (next to Log Out, iOS
only) so it is reachable on every device. ProfileIndex reuses the same
dialog instead of duplicating the confirm-by-username logic.
