+++ src/pages/profile/Index.vue
@@ -25,50 +25,13 @@
     </div>
     <locale-selector :visible="operating.locale == true" @close="operating.locale = false" />
     <help-dialog :visible="operating.help == true" @close="operating.help = false" />
-    <el-dialog
-      v-model="deleteDialogVisible"
-      :title="$t('common.nav.deleteAccount')"
-      width="90%"
-      align-center
-      class="delete-account-dialog"
-      @closed="onDeleteDialogClosed"
-    >
-      <el-alert
-        :title="$t('common.message.deleteAccountWarning')"
-        type="error"
-        :closable="false"
-        show-icon
-        class="mb-3"
-      />
-      <p class="delete-consequences">{{ $t('common.message.deleteAccountConsequences') }}</p>
-      <p class="delete-type-prompt">{{ $t('common.message.deleteAccountTypePrompt') }}</p>
-      <p class="delete-username">{{ user.username }}</p>
-      <el-input
-        v-model="deleteConfirmText"
-        :placeholder="$t('common.message.deleteAccountPlaceholder')"
-        autocomplete="off"
-      />
-      <template #footer>
-        <el-button round @click="deleteDialogVisible = false">
-          {{ $t('common.button.cancel') }}
-        </el-button>
-        <el-button
-          type="danger"
-          round
-          :disabled="!deleteConfirmMatched"
-          :loading="deleting"
-          @click="confirmDeleteAccount"
-        >
-          {{ $t('common.button.deletePermanently') }}
-        </el-button>
-      </template>
-    </el-dialog>
+    <delete-account-dialog v-model:visible="deleteDialogVisible" />
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import { ElImage, ElMessage, ElDialog, ElInput, ElButton, ElAlert } from 'element-plus';
+import { ElImage } from 'element-plus';
 import {
   ROUTE_CONSOLE_APPLICATION_LIST,
   ROUTE_CONSOLE_ORDER_LIST,
@@ -79,8 +42,8 @@ import {
 } from '@/router';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 import { getBaseUrlAuth, getBaseUrlPlatform, isIOS, withCurrentUserIdAndSite } from '@/utils';
-import { userOperator } from '@/operators';
 import HelpDialog from '@/components/common/HelpDialog.vue';
+import DeleteAccountDialog from '@/components/user/DeleteAccountDialog.vue';
 
 interface ILink {
   key: string;
@@ -95,11 +58,8 @@ export default defineComponent({
   name: 'ProfileIndex',
   components: {
     ElImage,
-    ElDialog,
-    ElInput,
-    ElButton,
-    ElAlert,
     HelpDialog,
+    DeleteAccountDialog,
     FontAwesomeIcon
   },
   data() {
@@ -109,20 +69,12 @@ export default defineComponent({
         dark: false,
         help: false
       },
-      deleteDialogVisible: false,
-      deleteConfirmText: '',
-      deleting: false
+      deleteDialogVisible: false
     };
   },
   computed: {
     user() {
       return this.$store.getters.user;
-    },
-    // Require the user to type their exact username to unlock the
-    // irreversible delete — a locale-independent guard against misclicks.
-    deleteConfirmMatched(): boolean {
-      const username = this.$store.getters.user?.username;
-      return !!username && this.deleteConfirmText.trim() === username;
     },
     showSupport() {
       return (
@@ -242,28 +194,7 @@ export default defineComponent({
       });
     },
     onDeleteAccount() {
-      this.deleteConfirmText = '';
       this.deleteDialogVisible = true;
-    },
-    async confirmDeleteAccount() {
-      if (!this.deleteConfirmMatched || this.deleting) {
-        return;
-      }
-      this.deleting = true;
-      try {
-        await userOperator.deleteMe();
-        this.deleting = false;
-        this.deleteDialogVisible = false;
-        ElMessage.success(this.$t('common.message.deleteAccountSuccess').toString());
-        await this.$store.dispatch('logout');
-        this.$router.push({ name: ROUTE_INDEX });
-      } catch {
-        this.deleting = false;
-        ElMessage.error(this.$t('common.message.deleteAccountFailed').toString());
-      }
-    },
-    onDeleteDialogClosed() {
-      this.deleteConfirmText = '';
     },
     onNavigate(link: ILink) {
       if (link.name) {
@@ -368,31 +299,6 @@ $padding-left: 10px;
     .text {
       font-size: 14px;
     }
-  }
-}
-</style>
-
-<style lang="scss" scoped>
-.delete-account-dialog {
-  .mb-3 {
-    margin-bottom: 12px;
-  }
-  .delete-consequences {
-    margin: 12px 0;
-    font-size: 13px;
-    line-height: 1.6;
-    color: var(--el-text-color-regular);
-  }
-  .delete-type-prompt {
-    margin: 12px 0 4px;
-    font-size: 13px;
-  }
-  .delete-username {
-    margin: 0 0 10px;
-    font-weight: 700;
-    font-size: 15px;
-    word-break: break-all;
-    color: var(--el-color-danger);
   }
 }
 </style>
